+++ examples/typescript-react-example/src/TransitiveType1.gen.tsx
@@ -6,9 +6,9 @@
 import * as TransitiveType1BS__Es6Import from './TransitiveType1.bs';
 const TransitiveType1BS: any = TransitiveType1BS__Es6Import;
 
-import {t2Alias as TransitiveType2_t2Alias} from './TransitiveType2.gen';
+import type {t2Alias as TransitiveType2_t2Alias} from './TransitiveType2.gen';
 
-import {t2 as TransitiveType2_t2} from './TransitiveType2.gen';
+import type {t2 as TransitiveType2_t2} from './TransitiveType2.gen';
 
 export const convert: (x:TransitiveType2_t2) => TransitiveType2_t2 = function (Arg1: any) {
   const result = TransitiveType1BS.convert((Arg1 == null ? undefined : Arg1));