+++ examples/typescript-react-example/src/TransitiveType2.gen.tsx
@@ -2,7 +2,7 @@
 /* eslint-disable import/first */
 
 
-import {t3 as TransitiveType3_t3} from './TransitiveType3.gen';
+import type {t3 as TransitiveType3_t3} from './TransitiveType3.gen';
 
 // tslint:disable-next-line:interface-over-type-literal
 export type t2 = (null | undefined | TransitiveType3_t3);