+++ Changes.md
@@ -1,5 +1,6 @@
 # master
 - Make ReactEvent.Mouse.t be a builtin type corresponding to `MouseEvent`.
+- Use `import type` syntax for TypeScript.
 
 # 3.48.0
 - Remove workaround in Flow that turns components into functions when exported.